Added segment tree. However, it cannot search, for example, element indexes. It is necessary to make a separate tree that will own the idempotent function, then it will be possible to find out the index of the desired element. Also, it does not know how to solve the problem of where exactly to give answers as indices (for example, the problem of finding the left maximum)

+++ Structures/NumberTheory/FloorLog.h
@@ -15,14 +15,14 @@ namespace NumberTheory {
 		static void initialize_to_or_clear_after(size_t n) {
 			floor_log.resize(n + 1);
 			if (maximum_computed_logarithm < n) {
-				for (int i = maximum_computed_logarithm + 1; i <= n; ++i) {
+				for (size_t i = maximum_computed_logarithm + 1; i <= n; ++i) {
 					floor_log[i] = floor_log[i / 2] + 1;
 				}
 			}
 			maximum_computed_logarithm = n;
 		}
 
-		static int get_floor_log(size_t n) {
+		static size_t get_floor_log(size_t n) {
 			FloorLog::initialize_to_or_clear_after(n);
 			return floor_log[n];
 		}
